backend_service/requirements.txt
fastapi==0.116.0
uvicorn==0.38.0
python-multipart==0.0.20
python-dotenv==1.2.1
pandas==2.3.2
google-genai==1.29.0
anywidget==0.9.7
ipywidgets==8.1.2
jupyterlab_widgets==3.0.15
traitlets>=5.14.1
numpy>=1.23.5
umap-learn>=0.5.5
hdbscan>=0.8.33
nltk>=3.8.1
pathos>=0.3.0
tiktoken>=0.7.0
tqdm>=4.66.1
yaspin>=3.0.2
requests>=2.28.2
numba>=0.60.0
protobuf>=4.25.2
PyYAML>=6.0
